Add JWT authorization and ownership checks for materials and upcycled products

+++ internal/materials/service.go
@@ -1,5 +1,7 @@
 package materials
 
+import "errors"
+
 func GetAllMaterials() []Material {
 	return FetchMaterials()
 }
@@ -9,10 +11,30 @@ func CreateMaterialService(material *Material) error {
 
 }
 
-func UpdateMaterialService(id string, material *Material) error {
+func UpdateMaterialService(id string, userID uint, material *Material) error {
+
+	existing, err := GetMaterialByID(id)
+	if err != nil {
+		return err
+	}
+
+	if existing.TraderID != userID {
+		return errors.New("you do not own this material")
+	}
+
 	return UpdateMaterialRepository(id, material)
 }
 
-func DeleteMaterialService(id string) error {
+func DeleteMaterialService(id string, userID uint) error {
+
+	existing, err := GetMaterialByID(id)
+	if err != nil {
+		return err
+	}
+
+	if existing.TraderID != userID {
+		return errors.New("you do not own this material")
+	}
+
 	return DeleteMaterialRepository(id)
 }
